refactor(mindmap): extract session-persistence.ts, add root error boundary
Item 7 slice 2 plus the root error boundary (item 9).

The launcher integration made the persistence split necessary rather than
merely tidy: PlatformBootstrap has to ask "is there real saved work?" before
mounting App, which meant the boot layer imported storage helpers from the
4,600-line component it renders. session-persistence.ts now owns the persisted
shape, the storage key, and load/write/summarize/clear. The migrations that
transform loaded data into live stores stay in App.tsx — they operate on the
running stores, not on storage.

AppErrorBoundary sits above the ready branch in PlatformBootstrap. Map.tsx
guards the canvas; nothing sat above App, so any render throw blanked the page.
It deliberately does not clear storage: a render crash must leave the saved
mindmap as intact as a token expiry does. Its fallback uses plain strings
because the locale provider lives inside the boundary.

Behaviour-neutral. 355 Vitest (3 new), typecheck, eval typecheck, build green.
